Minor refactor - move files and tidy imports
