+++ src/pages/members/[slug].tsx
@@ -3,14 +3,11 @@ import Link from 'next/link'
 import { Member } from 'generated/graphql'
 
 import Layout from 'components/layout/Layout'
-// import { Flex, Box, Stack, Heading, Tag, Avatar } from '@chakra-ui/core'
-//import { Flex, Tag, Avatar } from '@chakra-ui/core'
 import { Box, Heading, Image } from 'rebass'
 import Grid from 'components/common/Grid'
 import Text from 'components/common/Text'
-import Stack from 'components/chakraFix/Stack'
-import Tag from 'components/chakraFix/Tag'
-import { FaMapMarkerAlt } from 'react-icons/fa'
+import Stack from 'components/common/Stack'
+import Tag from 'components/common/Tag'
 import { useRouter } from 'next/router'
 import GET_MEMBER from 'graphql/get-member'
 import { useQuery } from '@apollo/react-hooks'
